tests/wpt/tests/html/canvas/offscreen/webgpu-access/2d.webgpuaccess.getTextureFormat.rgba16f.tentative.https.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.webgpuaccess.getTextureFormat.rgba16f.tentative.https</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.webgpuaccess.getTextureFormat.rgba16f.tentative.https</h1>
<p class="desc">getTextureFormat() returns RGBA16F for a float16 context</p>


<script>
var t = async_test("getTextureFormat() returns RGBA16F for a float16 context");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d', {colorSpace: "display-p3", pixelFormat: "float16"});

  _assertSame(ctx.getTextureFormat(), "rgba16float", "ctx.getTextureFormat()", "\"rgba16float\"");
  t.done();

});
</script>
